rename unconnectedHighways to islandshighways
