+++ index.js
@@ -12,6 +12,6 @@ module.exports = {
   'crossingHighways': require('./validators/crossingHighways'),
   'nodeEndingNearHighway': require('./validators/nodeEndingNearHighway'),
   'crossingWaterwaysHighways': require('./validators/crossingWaterwaysHighways'),
-  'unconnectedHighways': require('./validators/unconnectedHighways'),
+  'islandsHighways': require('./validators/islandsHighways'),
   'overlapHighways': require('./validators/overlapHighways')
 };
