+++ test/fixtures/islandshighways.mbtiles
